# api/docs/swagger.yaml
definitions:
  model.BoughtItemsList:
    properties:
      products:
        items:
          $ref: '#/definitions/model.Item'
        type: array
    type: object
  model.BuyItemRequest:
    properties:
      amount:
        type: integer
      product_id:
        type: string
      user_id:
        type: string
    type: object
  model.BuyItemResponse:
    properties:
      amount:
        type: integer
      message:
        type: string
      product_id:
        type: string
      product_name:
        type: string
      user_id:
        type: string
    type: object
...
  model.Item:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
  model.ListItems:
    properties:
      count:
        type: integer
      products:
        items:
          $ref: '#/definitions/model.Item'
        type: array
    type: object
...
  model.Status:
    properties:
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: This code is written by Nodirbek in third mont exam GOLANG
  title: Welcome to User-Product service
  version: "1.0"
paths:
...
  /v1/product/{page}/{limit}:
    get:
      consumes:
      - application/json
      description: get all products
      parameters:
      - description: page
        in: path
        name: page
        required: true
        type: string
      - description: limit
        in: path
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.ListItems'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: get all products
      tags:
      - Product
  /v1/product/buy:
    post:
      consumes:
      - application/json
      description: buy a product
      parameters:
      - description: Purchase a product
        in: body
        name: PurchaseInfo
        required: true
        schema:
          $ref: '#/definitions/model.BuyItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.BuyItemResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: buy a product
      tags:
      - Product
  /v1/product/create:
    post:
      consumes:
      - application/json
      description: Insert a new product with provided details
      parameters:
      - description: Create product
        in: body
        name: ProductDetails
        required: true
        schema:
          $ref: '#/definitions/model.Item'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Item'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: create product
      tags:
      - Product
  /v1/product/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete product
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Status'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: delete product
      tags:
      - Product
  /v1/product/get/{id}:
    get:
      consumes:
      - application/json
      description: Get product
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Item'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: get product by id
      tags:
      - Product
  /v1/product/update/{id}:
    put:
      consumes:
      - application/json
      description: Update ptoduct
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: Update Product
        in: body
        name: UserInfo
        required: true
        schema:
          $ref: '#/definitions/model.Item'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Item'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: update product
      tags:
      - Product
  /v1/products/get/{id}:
    get:
      consumes:
      - application/json
      description: get all purchased products by user id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.BoughtItemsList'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: get all purchased products by user id
      tags:
      - Product
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"